refactor: use snake_case utils index modules
