+++ glacium/cli/projects.py
@@ -6,7 +6,7 @@ from rich.console import Console
 from rich.table import Table
 from rich import box
 from pathlib import Path
-from glacium.utils.ProjectIndex import list_projects
+from glacium.utils.project_index import list_projects
 from glacium.utils.convergence.stats import execution_time, cl_cd_summary
