+++ glacium/recipes/fensap.py
@@ -1,7 +1,7 @@
 """Recipe containing jobs to run the FENSAP solver."""
 
 from glacium.managers.recipe_manager import RecipeManager, BaseRecipe
-from glacium.utils.JobIndex import JobFactory
+from glacium.utils.job_index import JobFactory
 from glacium.engines.fensap import FensapEngine
 
 @RecipeManager.register
