+++ glacium/recipes/grid_dependency.py
@@ -1,7 +1,7 @@
 """Recipe running grid dependencies and FENSAP."""
 
 from glacium.managers.recipe_manager import RecipeManager, BaseRecipe
-from glacium.utils.JobIndex import JobFactory
+from glacium.utils.job_index import JobFactory
 from glacium.engines.pointwise import PointwiseEngine
 from glacium.engines.fensap import FensapEngine
