+++ tests/test_job_factory_list.py
@@ -2,7 +2,7 @@ from pathlib import Path
 import sys
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 
-from glacium.utils.JobIndex import JobFactory
+from glacium.utils.job_index import JobFactory
 
 
 def test_analysis_jobs_registered():
