+++ tests/test_script_job_base.py
@@ -4,7 +4,6 @@ from pathlib import Path
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 
 from glacium.core.base import ScriptJobBase, EngineBase
-from glacium.utils.JobIndex import JobFactory
 from glacium.engines.engine_factory import EngineFactory
 from glacium.models.config import GlobalConfig
 from glacium.models.project import Project
